Content-Type header in HTTP custom endpoint response handlers are now respected (#1300)
Before the commit, adding a Content-Type header in the reponse from the handler of a http:request:foobar event would be ignored and set to text/plain because of using ctx.text to return the request. ctx.text will set Content-Type to text/plain.

Now, if a Content-Type header is detected it will use a plain ctx.body response which does not alter Content-Type.